Shifter functions and renamed them
