+++ src/background.js
@@ -1,5 +1,5 @@
 import './activity-logger/background'
-
+import importHistory from './browser-history/import-history'
 
 function openOverview() {
     browser.tabs.create({
@@ -9,7 +9,9 @@ function openOverview() {
 
 // Open the overview when the extension's button is clicked
 browser.browserAction.onClicked.addListener(() => {
-    openOverview()
+    console.log("Hello World");
+    //openOverview()
+    importHistory();
 })
 
 browser.commands.onCommand.addListener(command => {
